requirements.txt
torch>=1.10
torchvision>=0.11
opencv-python>=4.5
numpy>=1.21
Pillow>=8.0